requirements.txt
Django
gunicorn
psycopg2-binary
django-redis
djangorestframework==3.14.0

